examples/gpu: Avoid unnecessary casts

+++ examples/gpu/examples/blit2darray.go
@@ -219,12 +219,12 @@ func (e *Blit2DArray) Init(context *common.Context) error {
 
 	// set up texture data
 
-	imageSizeInBytes := imageWidth * imageHeight * 4
+	imageSizeInBytes := uint32(imageWidth * imageHeight * 4)
 
 	textureTransferBuffer, err := context.Device.CreateTransferBuffer(
 		&sdl.GPUTransferBufferCreateInfo{
 			Usage: sdl.GPU_TRANSFERBUFFERUSAGE_UPLOAD,
-			Size:  uint32(imageSizeInBytes * 2),
+			Size:  imageSizeInBytes * 2,
 		},
 	)
 	if err != nil {
@@ -290,7 +290,7 @@ func (e *Blit2DArray) Init(context *common.Context) error {
 
 	copyPass.UploadToGPUTexture(&sdl.GPUTextureTransferInfo{
 		TransferBuffer: textureTransferBuffer,
-		Offset:         uint32(imageSizeInBytes),
+		Offset:         imageSizeInBytes,
 	}, &sdl.GPUTextureRegion{
 		Texture: e.sourceTexture,
 		Layer:   1,
